+++ R/globals.R
@@ -1,3 +1,0 @@
-if (base::getRversion() >= "2.15.1") {
-  utils::globalVariables(c("rasterGrob","rows","string","citationStart","rasterGrob","uuid"))
-}