+++ R/phylopic_count.r
@@ -1,4 +1,0 @@
-#' Get the total number of PhyloPic images.
-#' 
-#' @keywords internal
-phylopic_count <- function() image_count()