+++ NEWS.md
@@ -1,3 +1,19 @@
+Version: 1.0.0
+==============
+
+* rphylopic has now been transfered to the palaeoverse community (new maintainer: William Gearty and author: Lewis Jones)
+* The package has been updated to work with PhyloPic API ver. >2.1.1
+* The package has been reworked to its core functionality of fetching silhouettes and plotting them in base R and ggplot2:
+  * get_uuid: this function enables users to get uuid(s) associated with a taxonomic name (new function)
+  * get_phylopic: this function enables users to get the PhyloPic silhouette associated with a specific uuid (replaces image_get)
+  * pick_phylopic: this function enables users to pick specific PhyloPic silhouettes when multiple are available for a given taxonomic name (new function)
+  * add_phylopic_base: this function retains it's core functionality and is used to add a silhouette to a base R plot (updated from add_phylopic_base)
+  * add_phylopic: this function retains it's core functionality and is used to add a silhouette to a ggplot2 (updated from add_phylopic)
+  * recolor_phylopic: this function retains it's core functionality and is used to recolor a silhouette (updated from recolor_phylopic)
+  * get_attribution: this function enables users to get the attribution data associated with a specific uuid (new function)
+  * save_phylopic: this function enables users to save PhyloPic silhouettes using various formats (replaces save_png)
+  
+
 Version: 0.3.4
 ==============